requirements.txt
konlpy
pandas
google-cloud-bigquery
scikit-learn
streamlit
WordCloud
google-oauth2-tool
db-dtypes
virtualenv
supabase
Pillow
numpy
matplotlib
regex
streamlit_folium
folium
time